Implement TAS for ILs and add Instant Race Restart (press Z).
- Frame Advance
- Checkpoints (Save States Alternative)
- Input Replay (simulates Keyboard Inputs)
- Frame Recording
- Interface for Brute Forcing
- Attempts to Determinize Time (also fix Animations)

Can easily be expanded to Multiplayer and All Modes, will work on that later...

Save States would be complicated to implement, previous attempt was buggy.
If not doing rendering, replaying everything until a certain Checkpoint is quite fast and this is done as alternative.
From my experience, Input Recording is not that useful, I just end up composing Tases Tick by Tick from the beginning anyway...

As a byproduct, this also adds Race Instant Restart to quickly retry a failed IL :D !

Signed-off-by: Pttn <28868425+Pttn@users.noreply.github.com>